fixed: small issues in newly implemented internal callbacks
Default function for `internal_calculate_size` is not checking if widget
has `internal_fit_layout` callback or not, and another issue is not
forwarding the `internal_calculate_size` callback to children when
the widget is not fluid.

Check for division by 0 in default function for `internal_relayout`
callback.

Implemented `smoll_context_initialize_layout` which calculates flex-box
based layout for UI tree.